Change gRPC client endpoint impl with endpoint changes
