+++ stdlib/grpc/src/main/ballerina/grpc/client_endpoint.bal
@@ -16,7 +16,7 @@
 
 # The gRPC client endpoint provides the capability for initiating contact with a remote gRPC service. The API it
 # provides includes functions to send request/error messages.
-public type Client object {
+public type Client client object {
 
     # Gets invoked to initialize the endpoint. During initialization, configurations provided through the `config`
     # record is used for endpoint initialization.
@@ -24,22 +24,43 @@ public type Client object {
     # + config - - The ClientEndpointConfig of the endpoint.
     public extern function init(ClientEndpointConfig config);
 
-    # Gets called every time a service attaches itself to this endpoint - also happens at module init time.
-    # Not supported in client endpoint.
+    # Calls when initializing client endpoint with service descriptor data extracted from proto file.
     #
-    # + serviceType - - The type of the service to be registered.
-    public extern function register(typedesc serviceType);
+    # + stubType - Service Stub type. possible values: blocking, nonblocking.
+    # + descriptorKey - Proto descriptor key. Key of proto descriptor.
+    # + descriptorMap - Proto descriptor map. descriptor map with all dependent descriptors.
+    # + return - Returns an error if encounters an error while initializing the stub, returns nill otherwise.
+    public extern function initStub(string stubType, string descriptorKey, map descriptorMap)
+                               returns error?;
 
-    # Starts the registered service. Not supported in client endpoint.
-    public extern function start();
-
-    # Stops the registered. Not supported in client endpoint.
-    public extern function stop();
-
-    # Returns the client connection which is used to send message to server.
+    # Calls when executing blocking call with gRPC service.
     #
-    # + return - - Client connection.
-    public extern function getCallerActions() returns GrpcClient;
+    # + methodID - Remote service method id.
+    # + payload - Request message. Message type varies with remote service method parameter.
+    # + headers - Optional headers parameter. Passes header value if needed. Default sets to nil.
+    # + return - Returns response message and headers if executes successfully, error otherwise.
+    public remote extern function blockingExecute(string methodID, any payload, Headers? headers = ())
+                               returns ((any, Headers)|error);
+
+    # Calls when executing non-blocking call with gRPC service.
+    #
+    # + methodID - Remote service method id.
+    # + payload - Request message. Message type varies with remote service method parameter..
+    # + listenerService - Call back listener service. This service listens the response message from service.
+    # + headers - Optional headers parameter. Passes header value if needed. Default sets to nil.
+    # + return - Returns an error if encounters an error while sending the request, returns nil otherwise.
+    public remote extern function nonBlockingExecute(string methodID, any payload, typedesc listenerService,
+                                              Headers? headers = ()) returns error?;
+
+
+    # Calls when executing streaming call with gRPC service.
+    #
+    # + methodID - Remote service method id.
+    # + listenerService - Call back listener service. This service listens the response message from service.
+    # + headers - Optional headers parameter. Passes header value if needed. Default sets to nil.
+    # + return - Returns client connection if executes successfully, error otherwise.
+    public remote extern function streamingExecute(string methodID, typedesc listenerService, Headers? headers = ())
+                               returns StreamingClient|error;
 };
 
 # Represents client endpoint configuration.
