+++ stdlib/grpc/src/main/java/org/ballerinalang/net/grpc/nativeimpl/client/BlockingExecute.java
@@ -15,102 +15,90 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.net.grpc.nativeimpl.servicestub;
+package org.ballerinalang.net.grpc.nativeimpl.client;
 
 import io.netty.handler.codec.http.HttpHeaders;
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.bre.bvm.CallableUnitCallback;
-import org.ballerinalang.connector.api.BLangConnectorSPIUtil;
-import org.ballerinalang.connector.api.Service;
-import org.ballerinalang.connector.api.Value;
-import org.ballerinalang.connector.impl.ValueImpl;
 import org.ballerinalang.model.types.TypeKind;
 import org.ballerinalang.model.types.TypeTags;
 import org.ballerinalang.model.values.BMap;
-import org.ballerinalang.model.values.BTypeDescValue;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
 import org.ballerinalang.natives.annotations.ReturnType;
+import org.ballerinalang.net.grpc.Message;
 import org.ballerinalang.net.grpc.MethodDescriptor;
-import org.ballerinalang.net.grpc.StreamObserver;
 import org.ballerinalang.net.grpc.exception.GrpcClientException;
-import org.ballerinalang.net.grpc.stubs.DefaultStreamObserver;
-import org.ballerinalang.net.grpc.stubs.NonBlockingStub;
+import org.ballerinalang.net.grpc.stubs.BlockingStub;
+import org.ballerinalang.net.http.DataContext;
 
 import java.util.Map;
 
 import static org.ballerinalang.bre.bvm.BLangVMErrors.STRUCT_GENERIC_ERROR;
-import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT;
-import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_END_POINT;
-import static org.ballerinalang.net.grpc.GrpcConstants.GRPC_CLIENT;
+import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_ENDPOINT_REF_INDEX;
+import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_ENDPOINT_TYPE;
 import static org.ballerinalang.net.grpc.GrpcConstants.MESSAGE_HEADERS;
 import static org.ballerinalang.net.grpc.GrpcConstants.METHOD_DESCRIPTORS;
 import static org.ballerinalang.net.grpc.GrpcConstants.ORG_NAME;
 import static org.ballerinalang.net.grpc.GrpcConstants.PROTOCOL_PACKAGE_GRPC;
 import static org.ballerinalang.net.grpc.GrpcConstants.PROTOCOL_STRUCT_PACKAGE_GRPC;
-import static org.ballerinalang.net.grpc.GrpcConstants.REQUEST_MESSAGE_DEFINITION;
-import static org.ballerinalang.net.grpc.GrpcConstants.REQUEST_SENDER;
 import static org.ballerinalang.net.grpc.GrpcConstants.SERVICE_STUB;
-import static org.ballerinalang.net.grpc.GrpcConstants.SERVICE_STUB_REF_INDEX;
 import static org.ballerinalang.util.BLangConstants.BALLERINA_BUILTIN_PKG;
 
 /**
- * {@code StreamingExecute} is the StreamingExecute action implementation of the gRPC Connector.
+ * {@code BlockingExecute} is the BlockingExecute action implementation of the gRPC Connector.
  *
  * @since 1.0.0
  */
 @BallerinaFunction(
         orgName = ORG_NAME,
         packageName = PROTOCOL_PACKAGE_GRPC,
-        functionName = "streamingExecute",
-        receiver = @Receiver(type = TypeKind.OBJECT, structType = SERVICE_STUB,
+        functionName = "blockingExecute",
+        receiver = @Receiver(type = TypeKind.OBJECT, structType = CLIENT_ENDPOINT_TYPE,
                 structPackage = PROTOCOL_STRUCT_PACKAGE_GRPC),
         args = {
                 @Argument(name = "methodID", type = TypeKind.STRING),
-                @Argument(name = "listenerService", type = TypeKind.TYPEDESC),
+                @Argument(name = "payload", type = TypeKind.ANY),
                 @Argument(name = "headers", type = TypeKind.OBJECT, structType = "Headers",
                         structPackage = PROTOCOL_STRUCT_PACKAGE_GRPC)
+            
         },
         returnType = {
-                @ReturnType(type = TypeKind.OBJECT, structType = CLIENT,
-                        structPackage = PROTOCOL_STRUCT_PACKAGE_GRPC),
+                @ReturnType(type = TypeKind.ANY),
                 @ReturnType(type = TypeKind.RECORD, structType = STRUCT_GENERIC_ERROR,
-                        structPackage = BALLERINA_BUILTIN_PKG),
-        },
+                        structPackage = BALLERINA_BUILTIN_PKG),        },
         isPublic = true
 )
-public class StreamingExecute extends AbstractExecute {
+public class BlockingExecute extends AbstractExecute {
     private static final int MESSAGE_HEADER_REF_INDEX = 2;
 
     @Override
     public void execute(Context context, CallableUnitCallback callback) {
-        BMap<String, BValue> serviceStub = (BMap<String, BValue>) context.getRefArgument(SERVICE_STUB_REF_INDEX);
-        if (serviceStub == null) {
-            notifyErrorReply(context, "Error while getting connector. gRPC Client connector " +
+        BMap<String, BValue> clientEndpoint = (BMap<String, BValue>) context.getRefArgument(CLIENT_ENDPOINT_REF_INDEX);
+        if (clientEndpoint == null) {
+            notifyErrorReply(context, "Error while getting connector. gRPC client connector " +
                     "is not initialized properly");
             callback.notifySuccess();
             return;
         }
         
-        Object connectionStub = serviceStub.getNativeData(SERVICE_STUB);
+        Object connectionStub = clientEndpoint.getNativeData(SERVICE_STUB);
         if (connectionStub == null) {
-            notifyErrorReply(context, "Error while getting connection stub. gRPC Client connector " +
-                    "is not initialized properly");
+            notifyErrorReply(context, "Error while getting connection stub. gRPC Client " +
+                    "connector is not initialized properly");
             callback.notifySuccess();
             return;
         }
-
         String methodName = context.getStringArgument(0);
         if (methodName == null) {
-            notifyErrorReply(context, "Error while processing the request. RPC endpoint doesn't " +
-                    "set properly");
+            notifyErrorReply(context, "Error while processing the request. RPC endpoint " +
+                    "doesn't set properly");
             callback.notifySuccess();
             return;
         }
-
-        Map<String, MethodDescriptor> methodDescriptors = (Map<String, MethodDescriptor>) serviceStub.getNativeData
+        Map<String, MethodDescriptor> methodDescriptors = (Map<String, MethodDescriptor>) clientEndpoint.getNativeData
                 (METHOD_DESCRIPTORS);
         if (methodDescriptors == null) {
             notifyErrorReply(context, "Error while processing the request. method descriptors " +
@@ -118,7 +106,7 @@ public class StreamingExecute extends AbstractExecute {
             callback.notifySuccess();
             return;
         }
-
+        
         com.google.protobuf.Descriptors.MethodDescriptor methodDescriptor = methodDescriptors.get(methodName) != null
                 ? methodDescriptors.get(methodName).getSchemaDescriptor() : null;
         if (methodDescriptor == null) {
@@ -126,62 +114,45 @@ public class StreamingExecute extends AbstractExecute {
             callback.notifySuccess();
             return;
         }
-
-        if (connectionStub instanceof NonBlockingStub) {
-            NonBlockingStub nonBlockingStub = (NonBlockingStub) connectionStub;
-
-            BTypeDescValue serviceType = (BTypeDescValue) context.getRefArgument(1);
-            Service callbackService = BLangConnectorSPIUtil.getServiceFromType(context.getProgramFile(), getTypeField
-                    (serviceType));
-
+        
+        if (connectionStub instanceof BlockingStub) {
+            BValue payloadBValue = context.getRefArgument(1);
+            Message requestMsg = new Message(methodDescriptor.getInputType().getName(), payloadBValue);
             // Update request headers when request headers exists in the context.
             BValue headerValues = context.getNullableRefArgument(MESSAGE_HEADER_REF_INDEX);
             HttpHeaders headers = null;
             if (headerValues != null && headerValues.getType().getTag() == TypeTags.OBJECT_TYPE_TAG) {
                 headers = (HttpHeaders) ((BMap<String, BValue>) headerValues).getNativeData(MESSAGE_HEADERS);
             }
-
+            if (headers != null) {
+                requestMsg.setHeaders(headers);
+            }
+            BlockingStub blockingStub = (BlockingStub) connectionStub;
             try {
                 MethodDescriptor.MethodType methodType = getMethodType(methodDescriptor);
-                DefaultStreamObserver responseObserver = new DefaultStreamObserver(callbackService);
-                StreamObserver requestSender;
-                if (methodType.equals(MethodDescriptor.MethodType.CLIENT_STREAMING)) {
-                    requestSender = nonBlockingStub.executeClientStreaming(headers, responseObserver,
-                            methodDescriptors.get(methodName));
-                    
-                } else if (methodType.equals(MethodDescriptor.MethodType.BIDI_STREAMING)) {
-                    requestSender = nonBlockingStub.executeBidiStreaming(headers, responseObserver, methodDescriptors
-                            .get(methodName));
+                if (methodType.equals(MethodDescriptor.MethodType.UNARY)) {
+
+                    DataContext dataContext = new DataContext(context, callback, null);
+                    blockingStub.executeUnary(requestMsg, methodDescriptors.get(methodName),
+                            dataContext);
                 } else {
                     notifyErrorReply(context, "Error while executing the client call. Method type " +
                             methodType.name() + " not supported");
-                    return;
+                    callback.notifySuccess();
                 }
-                BMap<String, BValue> connStruct = createStruct(context, GRPC_CLIENT);
-                connStruct.addNativeData(REQUEST_SENDER, requestSender);
-                connStruct.addNativeData(REQUEST_MESSAGE_DEFINITION, methodDescriptor
-                        .getInputType());
-                BMap<String, BValue> clientEndpoint = (BMap<String, BValue>) serviceStub.getNativeData
-                        (CLIENT_END_POINT);
-                clientEndpoint.addNativeData(GRPC_CLIENT, connStruct);
-                context.setReturnValues(clientEndpoint);
-                callback.notifySuccess();
             } catch (RuntimeException | GrpcClientException e) {
                 notifyErrorReply(context, "gRPC Client Connector Error :" + e.getMessage());
                 callback.notifySuccess();
             }
+        } else {
+            notifyErrorReply(context, "Error while processing the request message. Connection Sub " +
+                    "type not supported");
+            callback.notifySuccess();
         }
     }
 
     @Override
     public boolean isBlocking() {
         return false;
-    }
-
-    private Value getTypeField(BTypeDescValue refField) {
-        if (refField == null) {
-            return null;
-        }
-        return ValueImpl.createValue(refField);
     }
 }
