+++ stdlib/grpc/src/main/java/org/ballerinalang/net/grpc/nativeimpl/client/InitStub.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.net.grpc.nativeimpl.servicestub;
+package org.ballerinalang.net.grpc.nativeimpl.client;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
@@ -41,10 +41,10 @@ import java.util.Map;
 
 import static org.ballerinalang.bre.bvm.BLangVMErrors.STRUCT_GENERIC_ERROR;
 import static org.ballerinalang.net.grpc.GrpcConstants.BLOCKING_TYPE;
-import static org.ballerinalang.net.grpc.GrpcConstants.CALLER_ACTIONS;
+import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_CONNECTOR;
 import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_ENDPOINT_CONFIG;
 import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_ENDPOINT_REF_INDEX;
-import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_END_POINT;
+import static org.ballerinalang.net.grpc.GrpcConstants.CLIENT_ENDPOINT_TYPE;
 import static org.ballerinalang.net.grpc.GrpcConstants.DESCRIPTOR_KEY_STRING_INDEX;
 import static org.ballerinalang.net.grpc.GrpcConstants.DESCRIPTOR_MAP_REF_INDEX;
 import static org.ballerinalang.net.grpc.GrpcConstants.METHOD_DESCRIPTORS;
@@ -53,7 +53,6 @@ import static org.ballerinalang.net.grpc.GrpcConstants.ORG_NAME;
 import static org.ballerinalang.net.grpc.GrpcConstants.PROTOCOL_PACKAGE_GRPC;
 import static org.ballerinalang.net.grpc.GrpcConstants.PROTOCOL_STRUCT_PACKAGE_GRPC;
 import static org.ballerinalang.net.grpc.GrpcConstants.SERVICE_STUB;
-import static org.ballerinalang.net.grpc.GrpcConstants.SERVICE_STUB_REF_INDEX;
 import static org.ballerinalang.net.grpc.GrpcConstants.STUB_TYPE_STRING_INDEX;
 import static org.ballerinalang.util.BLangConstants.BALLERINA_BUILTIN_PKG;
 
@@ -66,11 +65,9 @@ import static org.ballerinalang.util.BLangConstants.BALLERINA_BUILTIN_PKG;
         orgName = ORG_NAME,
         packageName = PROTOCOL_PACKAGE_GRPC,
         functionName = "initStub",
-        receiver = @Receiver(type = TypeKind.OBJECT, structType = SERVICE_STUB,
+        receiver = @Receiver(type = TypeKind.OBJECT, structType = CLIENT_ENDPOINT_TYPE,
                 structPackage = PROTOCOL_STRUCT_PACKAGE_GRPC),
         args = {
-                @Argument(name = "clientEndpoint", type = TypeKind.OBJECT, structPackage = PROTOCOL_STRUCT_PACKAGE_GRPC,
-                        structType = "Client"),
                 @Argument(name = "stubType", type = TypeKind.STRING),
                 @Argument(name = "descriptorKey", type = TypeKind.STRING),
                 @Argument(name = "descriptorMap", type = TypeKind.MAP)
@@ -84,9 +81,8 @@ import static org.ballerinalang.util.BLangConstants.BALLERINA_BUILTIN_PKG;
 public class InitStub extends BlockingNativeCallableUnit {
     @Override
     public void execute(Context context) {
-        BMap<String, BValue> serviceStub = (BMap<String, BValue>) context.getRefArgument(SERVICE_STUB_REF_INDEX);
         BMap<String, BValue> clientEndpoint = (BMap<String, BValue>) context.getRefArgument(CLIENT_ENDPOINT_REF_INDEX);
-        HttpClientConnector clientConnector = (HttpClientConnector) clientEndpoint.getNativeData(CALLER_ACTIONS);
+        HttpClientConnector clientConnector = (HttpClientConnector) clientEndpoint.getNativeData(CLIENT_CONNECTOR);
         Struct endpointConfig = (Struct) clientEndpoint.getNativeData(CLIENT_ENDPOINT_CONFIG);
         String stubType = context.getStringArgument(STUB_TYPE_STRING_INDEX);
         String descriptorKey = context.getStringArgument(DESCRIPTOR_KEY_STRING_INDEX);
@@ -109,21 +105,19 @@ public class InitStub extends BlockingNativeCallableUnit {
             ServiceDefinition serviceDefinition = new ServiceDefinition(descriptorMap.get(descriptorKey).stringValue(),
                     descriptorMap);
             Map<String, MethodDescriptor> methodDescriptorMap = serviceDefinition.getMethodDescriptors(context);
-            
-            serviceStub.addNativeData(METHOD_DESCRIPTORS, methodDescriptorMap);
+
+            clientEndpoint.addNativeData(METHOD_DESCRIPTORS, methodDescriptorMap);
             if (BLOCKING_TYPE.equalsIgnoreCase(stubType)) {
                 BlockingStub blockingStub = new BlockingStub(clientConnector, endpointConfig);
-                serviceStub.addNativeData(SERVICE_STUB, blockingStub);
+                clientEndpoint.addNativeData(SERVICE_STUB, blockingStub);
             } else if (NON_BLOCKING_TYPE.equalsIgnoreCase(stubType)) {
                 NonBlockingStub nonBlockingStub = new NonBlockingStub(clientConnector, endpointConfig);
-                serviceStub.addNativeData(SERVICE_STUB, nonBlockingStub);
+                clientEndpoint.addNativeData(SERVICE_STUB, nonBlockingStub);
             } else {
                 context.setReturnValues(MessageUtils.getConnectorError(new StatusRuntimeException(Status
                         .fromCode(Status.Code.INTERNAL.toStatus().getCode()).withDescription("Error while " +
                                 "initializing connector. invalid connector type"))));
-                return;
             }
-            serviceStub.addNativeData(CLIENT_END_POINT, clientEndpoint);
         } catch (RuntimeException | GrpcClientException e) {
             context.setReturnValues(MessageUtils.getConnectorError(e));
         }
